backend_app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
paho-mqtt
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
